Payflow dotNET SDK v4.4.7
Add fields for Card on File (COF) - Stored Credentials - see https://developer.paypal.com/docs/payflow/integration-guide/new-features.

This will be the last build of this branch.   v5.0 will be released soon but is not 100% compatible with older builds.

For future use and to support any new NVPs you can always use the <EXTDATA> field to pass the name-value of the NVP you would like to support.